refactor: update Header, Hidden to vitest
