+++ packages/react/jest.config.js
@@ -73,6 +73,9 @@ module.exports = {
     '<rootDir>/src/internal/utils/',
     '<rootDir>/src/live-region/',
     '<rootDir>/src/utils/',
+
+    '<rootDir>/src/Header/',
+    '<rootDir>/src/Hidden/',
   ],
   transformIgnorePatterns: ['node_modules/(?!@github/[a-z-]+-element|@lit-labs/react|@oddbird/popover-polyfill)'],
 }
