+++ packages/react/src/Hidden/Hidden.test.tsx
@@ -1,7 +1,7 @@
 import {render} from '@testing-library/react'
-import {Hidden} from '.'
+import {describe, it, expect, afterAll, beforeAll} from 'vitest'
+import {Hidden} from '../Hidden'
 import MatchMediaMock from 'jest-matchmedia-mock'
-import {behavesAsComponent, checkExports, checkStoriesForAxeViolations} from '../utils/testing'
 
 let matchMedia: MatchMediaMock
 describe('Hidden', () => {
@@ -11,21 +11,6 @@ describe('Hidden', () => {
 
   afterAll(() => {
     matchMedia.clear()
-  })
-
-  behavesAsComponent({
-    Component: Hidden,
-    options: {skipAs: true, skipSx: true},
-    toRender: () => (
-      <Hidden when={'narrow'}>
-        <div>Hidden when narrow</div>
-      </Hidden>
-    ),
-  })
-
-  checkExports('Hidden', {
-    default: Hidden,
-    Hidden,
   })
 
   it('renders `when` prop as expected', () => {
@@ -65,6 +50,3 @@ describe('Hidden', () => {
     )
   })
 })
-
-checkStoriesForAxeViolations('Hidden.features', '../Hidden/')
-checkStoriesForAxeViolations('Hidden.examples', '../Hidden/')
